Reorganize batch renderer into a render batch. Add separate renderer class.
The batch renderer didn't feel quite right, so I instead made it just represent the data of a single
batch. The batch can then be passed to a renderer in order to be drawn. This means I don't need
lots of different renderers (for single objects, batches, and whatever else), and can prepare the
draw data more carefully.