Reorganize CMake build files to modularize and follow modern CMake guidelines
This was quite a restructruring and I'm sure something could be way better, but it compiles with the same result as before the refactor, and I now have three separate-ish modules. The external libraries should probably not be in the project the way they currently are, but I do not know how to fix that in a good way. Currently, it builds and works as before, just with more isolated chunks and subdirectories.